Don't wait for output messages
Currently, when part of our Hot Reload support wants to write to the "Hot Reload" output window pane it awaits `IHotReloadDiagnosticOutputService.WriteLineAsync`. The implementation of this interface needs to transition to the UI thread, and then it can access the `IVsOutputWindowPane` and write the message.

However, these messages are informational rather than functional; they give the user some idea of what Hot Reload is up to, but no part of Hot Reload actually depends on what gets written to the Output Window or when. As such there's no need for any of our Hot Reload operations to wait for the message to be output.

Here we make outputting these messages a "fire and forget" operation. The async `IHotReloadDiagnosticOutputService.WriteLineAsync` has become the sychronous `IHotReloadDiagnosticOutputService.WriteLine` method, but it just posts work to happen on the UI thread sometime later. All callers have been updated accordingly.