Optimize when fast up-to-date check is disabled
Previously when the fast up-to-date check was disabled we would still perform all project data ingestion, only to not use that data.

We would also allocate a new instance of `UpToDateCheckImplicitConfiguredInput` for each project configuration.

With this change we no longer perform data ingestion when disabled, and no longer allocate a data object each time the project updates.