Fix missing newline in output window
In recent code review, it was suggested to change from OutputString to OutputStringNoPump. My testing of that change only involved a single project, so the message printed was the last in the Output Window. However the no-pump version of this method does not include a trailing newline character, meaning that had there been another log message, it would have commenced on the same line that the incremental build failure warning finished on.

The change here makes our OutputStringNoPump extension method behave identically to OutputString in terms of what it writes to the output pane.