Handle missing launch profile
`LaunchProfilesDebugLaunchProvder.GetActiveProfileAsync` current throws an exception if there is no active launch profile. This is a problem, as there is _no_ guarantee that there will always be a launch profile.

We do try to provide a default launch profilewhen the launchSettings.json file doesn't exist (and this is normal when creating a new project--there's no reason to require a mostly empty launchSettings.json just so the user can run the project they just created with default settings). But that profile can be deleted or replaced by other components, and it may even be possible to ask for the launch settings before we've had a change to add that default.

So in the general case we need to handle the possibility that there are no launch profiles at all, and thus no active profile, rather than throwing an exception.

`GetActiveProfileAsync` is used by two other methods in the class for two different purposes. `CanBeStartupProjectAsync` is called to see if a project can be a startup project based on the launch profile. There's no reason to be throwing an exception here--we need a launch profile so we know how to start the project, and if we don't have an active launch profile then we definitely can't start the project.

On the other hande, `QueryDebugTargetsInternalAsync` is called when we have already decided to start/debug the project, and it is definitely exceptional that we're asked to do something that can't possibly be done. As such, the check and exception throw in `GetActiveProfileAsync` have moved here.

Currently the thrown exception from `CanBeStartupProjectAsync` is responsible for a large number of non-fatal exception reports, and this change should fix that.