Remove IWaitContext.AllowCancel
This property is unused and will not work correctly. If a dialog was created that did not allow cancellation, then this property is later set to true, the backing cancellation token will not be updated. Further, it's unlikely that a consumer of this type would re-query the token after setting the property. It feels like a safer design, and a cleaner implementation, to remove this unused property and avoid potential future bugs.