Notify operation progress even when we skip batches
In #7848 we introduced a mechanism to skip sending empty batches to the language service. That change introduced a bug where a skipped batch would also fail to provide operation progress with the data source versions.

This could result in a case where the last language service update was empty, we wouldn't notify operation progress of that update's version, and operation progress would run indefinitely.

This change ensures we always notify operation progress.

Note that this code is going to change in the next PR (already submitted as #7855). I will rebase that PR on this change once merged, and add unit tests to cover this scenario.