Expose implicit build state
Implementations of the `IImplicitlyTriggeredBuildManager` interface (defined in CPS) are called at the beginning and end of "implicit" builds--builds triggered as an incidental step in some larger operation such as F5, Ctrl+F5, debug unit test, etc. Our implementation of this, `ImplicitlyTriggeredBuildManager`, uses this to set a few specific global build properties. (Aside: these build properties were added to support skipping analyzers during implicit builds--and thus speeding them up--though they are no specific to that purpose.)

The issue with the current implementation is that the "implicitness" of a build is not exposed to anything other than the `ImplicitlyTriggeredBuildManager.GlobalPropertiesStore` type. We would like to add other global build property providers that can take it into account, specifically, one that limits the build of a multi-targeting project to a single target but only when that project is the startup project.

To support that we introduce a new interface, `IImplicitlyTriggeredBuildState`, and implement it in `ImplicitlyTriggeredBuildManager`. The manager now does nothing but indicate whether or not we are in an implicit build. The related `GlobalProviderProvider` type (aside: I think that's meant to be `GlobalPropertiesProvider` and I will update it in a future commit) now imports the new interface and uses it check if the build is implicit and return the correct set of properties from the `GlobalPropertiesStore`.