Remove assertion
This assertion seems logically sound but is failing in practice in some cases. I will investigate the reason for this in future. For now, it is harmless to continue in such cases. The only downside might be redundant batches being sent to Roslyn, but we had many more such updates only a week ago, so even if some of these slip through we are still in a much better state than before.