Don't require the UI thread for settings
Visual Studio settings are exposed through the `Microsoft.VisualStudio.Settings.ISettingsManager`. We hide this behind our `IProjectSystemOptions` interface, partly so we can access these options in the Microsoft.VisualStudio.ProjectSystem.Managed project (which doesn't depend on VS-specific interfaces). Currently, our implementation of `IProjectSystemOptions` treats `ISettingsManager` as being bound to the UI thread, which means it explicitly transitions to the UI thread before retrieving the service or calling methods on it.

However, this is unnecessarily restrictive, as the `ISettingsManager` documentation explicitly states that its methods may be called on any thread. Here we remove the transitions to the main thread, and access `ISettingsManager` through an `IVsService<T>` instead of `IVsUIService<T>`.

The immediate practical effect of this change is the Fast Up-to-Date Check will no longer need to transition to the main thread once (or more) for every project being built. This will also allow me to simplify the `ImplicitlyTriggeredDebugBuildManager`--it caches a certain setting value and listens for events to reset the value, presumably to avoid a transition to the main thread every time the value is needed. After this change I think we can get rid of all of that.