Make DebugType config-dependent
As with most of the other settings on the "Build" property page, the `DebugType` property is meant to be configuration-dependent. While the default for .NET Core projects is to always produce a portable PDB, this supports the common practice of producing a PDB for debug builds and no PDB for release builds.

We also need this property value to go through interception (see the `DebugTypeValueProvider` for details) and so we need to add a property-specific `DataSource` and set the `Persistence` to "ProjectFileWithInterception". When we add a property-specific `DataSource` we typically set the `HasConfigurationCondition` value again as we are no longer inheriting it from the `Rule`-level `DataSource` and the default value is often wrong.

In this case we made a mistake when initially creating BuildPropertyPage.xaml, and set `HasConfigurationCondition` to "False". This means the UI will not let you set different values in different configurations. The fix here is to change it to "True".