Track startup projects
The `ImplicitlyTriggeredDebugBuildManager` tracks when an "implicit" build is started as part of F5/Ctrl+F5 and notifies the `ImplicitlyTriggeredBuildManager`. Here we update it to also acquire and pass along the set of startup projects associated with the build (if any).

The existing `IStartupProjectHelper`/`StartupProjectHelper` already knows how to map from the VS understanding of the startup projects to the associated `UnconfiguredProject`s and retrieve a set of exports from those. This change adds a similar but simpler method to retrieve the full paths to the startup projects. The `ImplicitlyTriggeredDebugBuildManager` retrieves those full paths, and stores them via an `IImplicitlyTriggeredBuildManager2` interface--as the original interface is defined in CPS we can't simply add the method there.