Fast up-to-date learns about rebuilds
Fixes AB#1471474

This fixes a scenario where the first build after a rebuild can incorrectly be declared as not up-to-date.

The fast up-to-date check tracks the last time a project was built. However, `IBuildUpToDateCheckProvider` is only called for builds, not rebuilds. In order for the check to correctly update the last build time when a rebuild occurs, we introduce a new component that subscribes to SolutionBuildManager events and forwards them through to the relevant project.