Don't check configurations we aren't building
When a project targets multiple versions of the .NET framework (i.e. it has a `TargetFrameworks` property) CPS represents this as multiple `ConfiguredProject`s--one per framework. When the user builds (either through an explicit build or rebuild, or implicitly as part of F5, Ctrl+F5, running unit tests, etc.) CPS asks us to build _one_ of those specific configurations. This is generally not what we want, and so we force MSBuild to build for all the frameworks (this is accomplished by resetting the `TargetFramework` property to the empty string in the `TargetFrameworkGlobalBuildPropertyProvider` type).

Meanwhile, the Fast Up-to-Date Check feature supports this by examining _all_ relevant configurations--again, regardless of which one CPS requests we build. The inputs and outputs of each individual configuration are compared, and if any input is newer than any output the project as a whole is not considered up-to-date and we need to invoke MSBuild.

However, with the changes in #7884 we _don't_ always build for all the target frameworks. When we know which specific target framework the user is going to launch with F5/Ctrl+F5 we can often limit the build to just that single target framework (please see that PR for details as there are a bunch of conditions that must be met for this to work properly). This can speed up the code/build/debug development cycle by skipping the builds of the irrelevant frameworks.

The problem is that the Fast Up-to-Date Check is still checking all the configurations. After one of these "single target framework" builds, the Fast Up-to-Date Check will always report failure because the other configurations are out of date. So we'll only build one configuration, but we'll always try to build it even when we don't need to.

The fix here is to make the Fast Up-to-Date Check aware of these single target builds and restrict its checks accordingly. If a `TargetFramework` is specified in the global properties for the build, we need only consider the inputs and outputs for the corresponding configuration. Otherwise, we can skip the checks.

Note there are other ways of implementing this that might be, in some sense, cleaner than this approach. I can imagine updating CPS itself to be aware that we may want to build multiple configurations, and having some interface that allows it to run the Fast Up-to-Date Check individually for each of those configurations rather than assuming we want to build the single "active" configuration. However, that would be a lot of change for something that is straightforward to implement in our existing system.

Also, we have this overarching issue (the mismatch between the CPS concept of a single "active" configuration, and our need to work across multiple equally-important configurations) in other places (for example, the Dependencies node in Solution Explorer). If we are going to solve that problem, I would like us to solve it across the board and not just for one feature.