Optimize up-to-date check for CopyAlways items
Today, having a single `CopyToOutputDirectory` `Always` item in a project causes the fast up-to-date check to immediately flag the project as out of date, leading to a build.

It is very easy for users to select the `Always` option and unwittingly break incremental build performance.

This commit changes the default behaviour such that having an Always item doesn't fail the up-to-date check, Instead, we only fail if the files differ in size or timestamp.

Users who prefer the old behaviour can opt out of this optimization by setting the `DisableFastUpToDateCopyAlwaysOptimization` project property to true.

There were quite a few changes to the fast up-to-date check's log output to better explain behaviour in light of this change, such as including the `CopyToOutputDirectory` value for each item, along with item types.

Unit tests were expanded to cover these new behaviours.